refactor KottageList, invalidation on read
