+++ kottage/src/commonMain/kotlin/io/github/irgaly/kottage/internal/KottageStorageImpl.kt
@@ -203,7 +203,6 @@ internal class KottageStorageImpl(
         return KottageListImpl(
             name,
             this,
-            this.options.strategy,
             encoder,
             options,
             kottageOptions,
